fix for precission lost in large numbers

+++ src/main/java/com/jalvz/bio/depthcover/util/WriterUtils.java
@@ -9,7 +9,7 @@ public class WriterUtils {
 	
 	
 	public static String toStr(Number n) {
-		if (n.doubleValue() >= 0) {
+		if (n.floatValue() >= 0) {
 			return n.toString();
 		} else {
 			return "unknown";
